Manual

API

DensityFlows.AffineCouplingAxes — Type

AffineCouplingAxes(d, n=0; kws...)

Create axes for AffineCouplingLayer.

Arguments

d::Int: dimension of the flow.

n::Int: number of conditions / parameters (default is 0).

j::Int: dimension cut (default is d÷2).

reverse::Bool: (default is false).

The dimension cut j specifies which dimensions are not modified by the layer. If reverse is false the layer acts as the identity on dimensions (1, j). If reverse is true the layer acts as the identity on dimensions (j+1, d).

DensityFlows.AffineCouplingLayer — Type

AffineCouplingLayer(axes; kws...)
AffineCouplingLayer(d, n=0; kws...)

Create an AffineCouplingLayer with NN models s and t.

The layer can be represented as a function f such that on dimensions where it does not act like the identity it returns

\[ f(x) = x * \exp(s) + t \quad {\rm if~forward}\]

and

\[ f^{-1}(z) = \exp(-s) * (z-t) \quad {\rm if~backward} \, .\]

Arguments

axes::AffineCouplingAxes.

d::Int: dimension of the flow.

n::Int: number of conditions / parameters (default is 0).

j::Int: dimension cut (default is d÷2).

reverse::Bool: (default is false).

hidden_dim::Int: number of hidden dimensions in s and t (default is 32).

n_sublayers_t::Int: number of sublayers in t (default is 2).

n_sublayers_s::Int: number of sublayers in s (default is 2).

If axes is provided, then d, n, j and reverse should not be passed as arguments as they would be redondant.

The dimension cut j specifies which dimensions are not modified by the layer. If reverse is false the layer acts as the identity on dimensions (1, j). If reverse is true the layer acts as the identity on dimensions (j+1, d).

See also AffineCouplingAxes.

DensityFlows.AffineCouplingBlock — Type

AffineCouplingBlock(axes; kws...)

Create an block of two AffineCouplinglayer with opposite axes.

Opposite axes here means that one is set with reverse = true and the other with reverse = false.

Arguments

axes::AffineCouplingAxes.

hidden_dim::Int: number of hidden dimensions in s and t (default is 32).

n_sublayers_t::Int: number of sublayers in t (default is 2).

n_sublayers_s::Int: number of sublayers in s (default is 2).

See also AffineCouplingAxes and AffineCouplingLayer.